fix(company data): fix dumplicate issue
